feat: dissolve admin levels
Makes sure that the admin data is saved dissolved to a single polygon
per country. Additionally, a check is added when deciding on which
admin_level to use, make sure that if both 4 and 2 are available, 4
should only be chosen when the area of 4 is greater than that of 2. If
this is not the case, then the admin_level 2 is incomplete.